feat!: don't split tsconfig.dom.json into two files
I found it not very useful to split the dom config into two files.

- Even `@vue/compiler-dom` and `@vue/runtime-dom` are not split into
  different packages. They have made the `-core` versions their
  dependencies.
- The DOM-specific options are too generic. They are not specific to
  Vue, but to the browser environment. If separated from the base config,
  I'm afraid it doesn't deserve the `@vue/` scope. It's more useful when
  it automatically includes the Vue base config.

On the other hand, this change makes the migration from earlier versions more
complicated.

I'm sorry for the inconvenience brought to the early adopters. But
reverting it now is better than later.

Luckily, I haven't incorporated this change into `create-vue` yet.
According to the npm download stats, the downloads of v0.2.0 is about
1% of v0.1.3. I hope I'm reverting it early enough.