feat!: add `jsxImportSource`; Vue 3.3+ is required
BREAKING CHANGE: drop support for Vue 3.2 and below

See https://blog.vuejs.org/posts/vue-3-3#jsx-import-source-support

It doesn't need to be explicitly set for most projects because Volar /
vue-tsc would detect the Vue version and set it automatically.

But in a pure TS/TSX project, it's required to set it manually. So we
are explicitly adding it here.