chore: remove the FIXME comment on `outDir` field
Starting in TypeScript 5.6, `.tsbuildinfo` file is always written in a
`--build` invocation[^1].
So the entire community need to embrace the new behavior and explicitly
set `tsbuildInfoFile` if they want to avoid polluting the project root.

So I think it's no longer our responsibility to find a workaround for
the users.

Closes #27

[^1]: https://devblogs.microsoft.com/typescript/announcing-typescript-5-6-beta/#tsbuildinfo-is-always-written