fix: require vue 3.3 as peer dependency
Because `jsxImportSource: "vue"` is only available in Vue 3.3 and above:
<https://blog.vuejs.org/posts/vue-3-3#jsx-import-source-support>

I also made both `vue` and `typescript` optional peer dependencies,
in case the configuration is not used to *run* or *type-check* a vue
project, but only to *transpile* some source code.

Closes #17, as we never intended to support Vue 2.x in the first place,
even it works in some earlier versions, the configuration might not be
optimal for Vue 2.x projects.