feat!: turn on noUncheckedIndexedAccess for libraries
See https://www.semver-ts.org/formal-spec/5-compiler-considerations.html#strictness
for the arguments for turning on noUncheckedIndexedAccess for libraries.

In short, it helps prevent shipping accidental breaking type changes to
library consumers.

(Note the link also suggests turning on `exactOptionalPropertyTypes`,
but I still have doubts about it, as it would not be helpful without
a specific coding style, which is hard to enforce.)